Use input type=date
The reason I didn't originally is because it uses the American date
formats by default. I fixed that by doing:

	export LANG=en_IE.UTF-8

The way this all works is ... not great: just setting LC_TIME doesn't
work in Firefox, and in Chromium it still shows me American dates :-/
There's probably something I to tweak there, too.

Especially on mobile this is a much better UX though.