Make scale numbers clearer
Move to right, because I find that's where my eye looks for it. Also
don't use the grey colour and set it as bold, so it's more noticeable.

Fixes #21; I tried using different colours and such, but I didn't like
the results much and it got rather complex rather fast. Let's see how
well this works in practice.