UTC everywhere
Always show everything in UTC; this was already how it worked on
goatcounter.com, but do it on other computers too (my computer's TZ is
now set to UTC +8 and everything looked funky (depending on the time).

Should add feature to show stuff in local TZ, but it's tricky to do that
right. This is okay for now (at least it's consistent).