Don't prefill blank days in hit_stats
Right now every day for every path *must* have an entry since the site
start, even if there are no hits that day.

This is actually a lot of data, currently 640k rows/101M:

	goatcounter=# vacuum full;
	VACUUM

	goatcounter=> select count(*) from hit_stats;
	count
	--------
	637240

	goatcounter=# \d+
	Schema |      Name      |   Type   | Owner  |    Size    | Description
	public | hit_stats      | table    | martin | 101 MB     |

Deleting the empty ones reduces this to 40k/4M:

	goatcounter=# delete from hit_stats where stats='[0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]';
	DELETE 604659

	goatcounter=# vacuum full;
	VACUUM

	goatcounter=# \d+
							List of relations
	Schema |      Name      |   Type   | Owner  |    Size    | Description
	--------+----------------+----------+--------+------------+-------------
	public | hit_stats      | table    | martin | 3992 kB    |

Doing this at runtime is faster than I would have expected, for the last 6
months on arp242.net it *was*

	HitStats.List   293ms  select hits
	HitStats.List    83ms  select hits_stats
	HitStats.List   111ms  reorder data
	HitStats.List   122ms  get total

And is now:

	HitStats.List   137ms  select hits
	HitStats.List    24ms  select hits_stats
	HitStats.List    74ms  reorder data
	HitStats.List     3ms  fill blanks
	HitStats.List    88ms  get total

Only 3ms to fill in the blank data, but we save 249ms in doing everything else!