Disable keyboard input on Pika datepicker
Disable keyboardInput as <Backspace> cleans the entire field instead of
just one character, which is unexpected and from the error logs many
people get this wrong (they submit stuff with a date like "8" because
they press "<BS>8<CR>").

This should be fixed better though, this is just a hotfix. Something for
the morning.