Link to paths by allowing people to fill in domain
Remove hits.domain, as most people use GoatCounter from just one domain,
so just use a global domain default.

There are some cases where this isn't the case and people use it from
different subdomains, but I'm not sure if there's an easy way to support
that in the current UI: to get paths to show up individually you need to
frob with the paths anyway, so they won't be accurate.

I'm not sure if there are a lot of people using it like this, thus far
only one, and they had to hack with Google Analytics as well.

Can always revisit based on feedback/requests, instead of adding
something complex that (almost) no one uses/needs anyway.

Additionally, we also need to separate out "paths" and "events" for
this, as we can't link to events. I'll document/implement better support
for this later, before the 1.1 release.