Support -tls parameter
Basic TLS support based on .pem files. I looked at doing ACME, but it's
rather complicated as this requires dns-01 verification, which requires
setting a "_acme-challenge.domain.com" record.

https://github.com/go-acme/lego includes support for this, but it still
requires giving lego (and thus goatcounter) API access to the domain,
which is rather meh.

I solved this on goatcounter.com using CNAME to an unimportant domain:

- goatcounter.com is managed by TransIP.

- goatcount.com is managed by CloudFlare (it's the only domain on my
  account there)

- _acme-challenge.goatcounter.com is a CNAME to goatcounter.goatcount.com

- In the dns-01 verification we set the goatcount.com address, so we
  only need to have the cloudflare API keys.

I think the best solution is to implement #120 first. With that, we
don't need to muck about with wildcard certs, and can just use http-01
challenge for one domain.