Remove touch events from drag_timeframe
This made scrolling the page very hard, since every time your finger is
on one of the charts it would start the dragging.

Not sure what a good fix is (only start if we move more than n pixels to
left/right?), and just disable for now: not a very critical feature.