Show banner reminding people about financials for serve and saas commands
I thought about this for a while: on one hand I'm not a fan of
displaying "noise" in the terminal, but OTOH I don't think it's too
unreasonable to remind people of this. And it's not like yo're
starting/stopping the server all the time, so displaying this just for
serve and saas seems okay.