Remove /set-tz; use json type for site.settings; don't drop table
Most people got a TZ set now, so no need to run /set-tz.

The JSON datatype doesn't really make a different for the app, but
makes it easier to query.

There's no real reason to drop tables in the schema files, as it's
intended just for creating new databases, and people might lose data if
they re-run it accidentally or some such. Better to just error out.