Don't display total pageviews until it's correct, and use same pagination logic
The "total displayed" is correct, but "total" is displayed in UTC, not
the configured timezone. I want to look at some DB issues first, and
probably won't fix it in the next few days. It's better to not display
data than display wrong data.

Also use the same pagination logic instead of relying on total >
totalDisplay.