Add GoatCounter script on the GoatCounter backend
Mainly just to see which browsers people are using and how many people
use mobile, so it's easier to make informed decisions about this.

Also hash the site.Code in the noscript tracking; I don't really need to
know *who* isn't using JS, just the total number of users (I can just
calculate the hashes on all the sitecodes, so not really "secure"; it's
just so I don't have an overview of it on the dashboard).