Don't infinitely loop when trying to display sites before the site creation
The startdate is capped to the site's creation date, and the loop would start
with a date after the end date, so this condition:

    if dayFmt == endFmt {

was never true.

I forgot why I didn't use time.After() in the first place; it may be due to TZ
offset issues 🤔 Need to check later, but that would be a (potentially) less
severe bug than this.

Also set the Deadline() lower; it was only 10 seconds temporarily while I was
fixing some performance issues for very large sites (which is now done). Can
probably be a bit lower still, but need to check.