Store paths and user_agents in their own tables
This changes how the paths and user_agents are stored: instead of
duplicating it for every single pageview, store it in their own table.
This saves quite a bit of space and improves query performance across
the board.

The database migration for this is a bit of a beast, and may take a
while if you have many (millions) of pageviews.

TODO:

- Update the code.
- Write SQLite migration.