Change default -listen to :443, document -listen and -tls in detail
The default -listen value of localhost:8081 probably isn't a good
default for many people: it kind of assumes that you're running a proxy
in front of it, and you need to play some tricks to get ACME to work
well.

So change it to :443, and also document the entire shebang in more
detail in "goatcounter help listen". ACME makes things a bit tricky
here, but hopefully this should clear it up for most people.