Drop GPL compatibility from license
It's quite easy to circumvent the "network protection copyleft" by
making some changes and then redistributing the entire work as the GPL,
MPL, or any of the other "weaker" copyleft licenses.

This isn't great, and rather defeats the purpose of having the stronger
copyleft in the first place. So reduce the "Compatible License" appendix
to licenses with similar copyleft as the EUPL has.