blacklist → refspam
To be honest I don't really buy the whole "blacklist is racist" thing,
but things generally seem to be moving away from "Blacklist", so makes
some pragmatical sense to do it here too. Will avoid potential and
pointless accusations/arguments in the future.

Besides, "refspam" is probably clearer here anyway, as we may have other
"blacklists" in the future.