Restyle tooltips from black background to grey background
I never really liked the black background, and used it just because
that's what Firefox and Chrome use by default. It's often hard to read
on my screen due to the different contrast. Granted, my screen is kinda
crap, but GoatCounter should work well with crap screens too.

Firefox/Chrome get around this by making it very large. I tried that,
and it looks ugly, so just restyle it to a light background to fit the
rest of the theme.