Rename "personalplus" plan to "starter"; limit data retention
A few changes to the plans: rename "personalplus" to "starter" and allow
commercial use on it. Below is email I'll send out to people later on
this, which also works as a commit message:

Note: the data retention isn't implemented yet; this is pretty easy to
add later, and adding it now means writing exceptions for existing
sites. Easier to just implement it later.

---

Hi there!

I made some changes to the GoatCounter plans/pricing:

- GoatCounter now has a "Starter" plan, this is €5/month, limited to 200k
  pageviews/month, comes with a custom domain, and allow commercial use.
  This is mostly the same as the "personal plus" plan there was before, except
  that it allows commercial use. If you had a "personal plus" for a custom
  domain before then you now have a Starter plan.

- Starting on August 1st the data retention will be limited for the Free and
  Starter plans: the Free plan will be limited to 6 months, the Starter plan to
  12 months, and the business plans remain unlimited.

  There is an export feature if you wish to retain your old pageviews.

Some background on this:

There seems to be a gap between "free for personal use" and "€15/month for
commercial use". I've gotten quite a bit of feedback of small (potential)
commercial users who just run a small website, where €15/month really is
prohibitively expensive.

The entire idea behind making it free for personal use is that I'd like
GoatCounter to be usable by as many people as possible, while also ensuring
commercial users pay their fair share. Redistributing software is free, but
developing it is not.

The general thinking is that larger businesses with several employees (who can
easily afford €15/month) will have more than 200k pageviews/month, whereas for
most startups and the like 200k should be more than enough.

The entire thing is a bit of a balancing act 😅 I may tweak the pricing further
in the future based on additional experience and feedback.

As for the data retention: the biggest issue here is that some not-
especially-active sites have had short bursts of *millions* of pageviews in a
short time because they wrote or made something that got widely shared.

The hits/months limit isn't strictly enforced because I don't want to tell
people to get a plan just because they wrote a popular article that got to the
front page of HN, Reddit, Twitter, etc, and GoatCounter has no problem handling
these levels of pageviews, so that's all fine.

But on the other hand, a million pageviews currently takes up about 400M of disk
space including backups (although this could probably be reduced a bit with a
more clever backup strategy). Disk space is pretty cheap, but it does add up.

It also means more effort on scaling GoatCounter; limiting the data retention is
an easy way to reduce the pressure on this. It also gives people a bit more
incentive to get a plan 😅

As always, self-hosting isn't affected by any of this. This just applies to the
goatcounter.com service.

Feel free to let me know if you've got any questions or feedback.

Cheers,
Martin