Set Cache-Control headers for static files in goatcounter serve
The /count.js has a short expiry of just a day, this way people will get
a cached version when navigating the page, but you won't run in to
problems if count.js adds a new feature and people still use an old
version which doesn't send the new data (resulting in wrong/outdated
data).

The rest will be cached for 30 days; the static files in the backend are
served with a "cache busting" query parameter:

    http://new.arp242.net:8081/all.min.css?v=95da6e2_2020-07-16T21%3a44%3a33%2b08%3a00

So a longer cache should be fine. Note that the default version of this
is "dev", and you must set it at compile-time, so update the README and
mention in CHANGELOG that people need to do this.