Store sessions in memory, instead of the database
Gives some decent performance gains, is actually a lot simpler, and not
storing any data (even temporarily) is probably better.

Some people are probably still going to say "but you're still collecting
personal information with your random IP! Storing it in memory is just
the same!" ... Well, okay, but this applies to any networked
application, since they all store your IP in memory.

The biggest downside is that it won't scale to multiple GoatCounter
instances since the sessionID is unique only to the current instance.
Need to think what to best do about that.