Cache hit_counts and ref_counts in memory for updating
Otherwise existing* queries run every time we persist hits to the DB for
all combinations; while the query is pretty fast, they get run *a lot*;
for example:

    Calls     Total   Mean    Min     Max     StdDev  Cache   Query
    633,275   1.5min  0.1ms   0.0ms   325.1ms 1.7     100%    existingRefCounts

That's since the app was restarted ~42 hours ago; about 220/minute. The
other queries have similar stats. This affects the performance of the
dashboard as well.

We fall back to a DB check to load the correct data on startup. This can
be improved further by just doing this on startup once, and then we
never need to run that afterwards.

The other update functions can also use this, but that's a bit more work
(the *_count already use "upsert" rather than "delete+insert"). Let's
add it here first an see how it goes.

The go-cache library seems okay; although it's not very actively
maintained there don't seem to be serious outstanding bugs and it's nice
and simply. We don't really have concurrency issues here since this is
essentially just a "for .. { .. }" loop; we could probably do with an
unsynced map as well; it's just more convenient for the cache
expiration, and the syncing should be a tiny performance hit here.

This also won't work with multiple GoatCounter instances; we should
abstract to a generic cache interface which can use memory or memcached,
Redis, or whatnot. This is already an issue for the sessionPaths cache
in Memstore (which, by the way, could also use this).

Other caches:
https://github.com/allegro/bigcache
https://github.com/coocood/freecache
https://github.com/dgraph-io/ristretto
https://dgraph.io/blog/post/caching-in-go/