Fix a bug with local subsite links.
Currently subsite links don't work when self-hosting. This is
because the template includes a leading `//` before the full
`site.URL()`, which already includes the site protocol. This
results in a link to something like `http://http://hostname`,
which the browser has trouble with.

This fixes the bug by simply removing the `//` in the template,
relying on `site.URL()` to be a fully qualified and navigable
URL that includes the protocol.

Let me know if there's UI tests in place that I should extend
as part of this.