Calculate browser etc. stats over UTC instead of site TZ
These are stored per-day, rather than per-hour. So if the site's TZ is
set to UTC +8 then the totals would be the totals for this TZ. In some
cases this skews the stats a bit as the percentages are calculated over
the wrong totals. This is especially an issue when viewing short time
periods and/or when there are a large number of pageviews just outside
of the selected range.

This still isn't "100% correct" as such, but is "correct enough" for
most purposes, and how it was done before as well, and no one complained
about that 🙃 A better solution would be to store it per-hour (and also
store the path while we're at it, so we can filter appropriately), but
that's a much larger refactor.

Ref: #337 (doesn't fix the location stats, which are still somewhat off
for some reason)