Reindex by month, rather than by day
Uses more memory, but is a bit faster.

Before (for arp242.net):

    Elapsed (wall clock) time (h:mm:ss or m:ss): 3:59.23
    Maximum resident set size (kbytes): 131064
    Minor (reclaiming a frame) page faults: 19957

After:

    Elapsed (wall clock) time (h:mm:ss or m:ss): 3:02.50
    Maximum resident set size (kbytes): 251176
    Minor (reclaiming a frame) page faults: 24608

Still want to make it faster, as 3 minutes times 2000 sits is about an
hour and a half. Most of the time is spent in the hit_counts and
ref_counts updates, as the queries in the existing* functions for those
are run so often as they're grouped by hour.