Allow concurrent connections to SQLite again
(actual change is in zgo.at/zdb)

The reason I disabled this is because the old design of sessions would
write to the database on every /count; this could cause some locking
issues with SQLite, and has long since been changed: this was a bad
design because 1) it didn't actually need to access the database, and 2)
the entire point of "memstore" was that most database writes were
limited to one thread.

Right now most database writes to hits are limited to memstore, which is
single-threaded, and writes to the *stats tables is done single-threaded
by cron. There are a few exceptions to this (i.e. the "purge" feature,
deleting sites), but they shouldn't cause that many issues in my
testing.

The journal_mode=wal (changed at the same time when the connections were
limited) greatly improves SQLite concurrency support.

I wrote a "writeOnceDB" wrapper around zdb.DB which ensures only one
connection is allowed to write simultaneously, but I don't think it's
needed, so let's not use that for now.

Limiting the connections didn't use to matter much in terms of
performance, but since a few months it gets the data from the dashboard
in ~8 parallel goroutines, greatly speeding things up, but not if there
is just one database connection to read stuff 🙃