Add autofocus to email and TOTP fields
This will help typing in these fields as it's not necessary
to manually focus them before typing.

Also, since (at least my) password manager copies the TOTP code to
the clipboard when autofilling entries, this has the benefit
of also not having to manually focus the TOTP field before pasting.