Pass site_id as meta data to Stripe, and store the billing_anchor
For some reason the checkout.session.completed webhook isn't always
sent; I also see this in testing: there's no webhook in the Stripe
dashboard either.

As a workaround, pass the site_id as metadata for the subscription; we
can then use that. This is also more convenient in the Stripe dashboard
anyway.

Also store the billing_anchor so we can calculate when the next invoice
will be without making API calls to Stripe. Will be used in the future.