Don't pass arguments to handler functions from mount
The problem here is that these variables get created once on startup
(when the handler is registered), and will remain the same afterwards.

A simpler example:

    func main() {
	x := f(0)
	x(false)
	x(true)
	x(false)

    }

    func f(n int) func(bool) {
	return func(edit bool) {
	    if edit {
		n = 42
	    }
	    fmt.Printf("%p %d\n", &n, n)
	}
    }

Will output:

    0xc00010c278 0
    0xc00010c278 42
    0xc00010c278 42

This was especially a problem in the editUser case:

    Get("/settings/users/{id}", h.usersForm(nil, nil))

As it could load the passed newUser from the URL, meaning that this
effectively became:

    Get("/settings/users/{id}", h.usersForm(&User{ID: 42, ...}, nil))

And people would see the wrong user 😬

The "one function to render the form"-pattern is pretty useful, but I
need to look at a way to do this better. For the time being just wrap
everything in a function, which works but is a bit ugly, as doing a more
serious refactor will conflict with some other branches I'm working on
and no one likes spending ages solving merge conflicts.

For most other cases it's not really an issue as the variable never gets
modified, but wrap these too just in case.