Filter "ignore IP" and bots in /api/v0/count
The API will now filter "Ignore IPs" if this is set; this can be turned
off with the Filter setting in the API request (defaults to enabled).

Filtering bots should probably just always happen; I considered adding a
value to Filter for this, but I'm not sure if this is really needed.

Fixes #476