Deal with visibilityState better
visibilityState=hidden is issues if someone opens a page in a background
tab, and this view will never be counted.

So instead we hook in the visibilitychange event and count the pageview
the first time the page becomes "visible". This seems supported by most
browsers:
https://caniuse.com/pagevisibility
https://caniuse.com/mdn-api_document_visibilitychange_event

If visibilityState isn't supported then we just count the request. There
are a few small but still used browsers for which this is the case, such
as Opera Mini.

Fixes #487