Add loading=eager
In Chrome "lite mode" the default loading mode is set to "lazy"; this is
acually kind of nice since only content that will be rendered on the
screen will be loaded.

However, for the tracking pixel we don't want this: we want to load it
right away. So explicitly set loading=eager to use the standard HTML
behaviour.

Fixes #505