Ignore unknown query parameters in /count
Not sure if there's ever a good reason to error out on it? Only downside
is that accidentally using the wrong parameter won't give an error now
but will be silently ignored, but I think that's okay – it's rare people
write their own integrations for this.

In response to the use case:

> I'd like to add ngsw-bypass so angular's service worker doesn't do
> horrible things to the request when it takes over the page.