Add web UI to translate GoatCounter
The web UI should really be part of zgo.at/z18n, but it's a lot easier
to build/prototype in here for now.

We store all user-edited translations in the "store" table per-user, and
do some shizzle to override the default translations with this. This is
so that you can preview translations you're working on without bothering
anyone else, which is kinda nice.