Get correct month/day names for timezones with negative TZ offset
Date.UTC(2017, 1, 1, 0, 0, 0) will return the "Dec 31 2016 16:00:00
GMT-0800" if the system's TZ is set to America/Los_Angeles. It doesn't
return  a "UTC time" but "local time in UTC".

Just use the local time; that should be fine.

Also don't need to translate anything if the system is already in
English.

Fixes #532 (again)