Reïmplement the charts with <canvas> and add line charts
The current bar charts served us well enough, but were really a hack
I've been intending to replace for a long time. There are a few problems
with the current approach:

- There are many elements to lay out; a week is 24 * 7 = 168 divs, times
  11 for the default view = 1848 divs. With a month it's 24 * 30 * 11 =
  7920 divs. This is quite slow, and really noticeable especially on
  older devices.

- The filesize is quite large: a week is ~100k, two months ~770k. While
  this compresses quite well (down to 85k for 2 months) because so much
  is repeated text, it's still pretty large.

  For reference, with this a week is 50k, and 2 months is 180k, which
  compresses to 29k.

A big reason I did it like this is because initially I wanted
GoatCounter to work reasonably well without JavaScript, but I've long
since let that go, and we have the "text table" view now too for JS-less
people and/or screenreaders.

I looked at a whole bunch of charting libraries and such, and many are
slow, large, or both. Turns out that just writing ~200-300 lines of code
to draw on a canvas works fine. uPlot[1] was pointed out[2] after I already
finished this implementation, and some quick evaluation/testing shows
it's quite fast, and small. Overall it's quite nice, but I'll stick with
what I have for now.

This adds both "bar charts" which look similar to the old ones, and a
new "line chart" which has a bit of a sleeker look. I set the line chart
as the default, including for existing users. I think that's what most
people want. You can switch in the dashboard settings.

This also moves the "view as text table" to the dashboard settings; the
checkbox in the dashboard header predates the dashboard settings page.
It does preserve the setting for people who set this as a default view,
because if you set it to be a text table then you probably want to
preserve that.

I dropped the "stacked" charts which shows both pageviews and visits;
although number of visits is still visible in the tooltip for now. I
might end up removing that, too. I don't think it's actually all that
useful, and if you don't want to count "visits" you can turn it off in
the "data collection" settings.

[1]: https://github.com/leeoniya/uPlot
[2]: https://github.com/arp242/goatcounter/pull/540#issuecomment-988593677