Add "server mangement" tab in settings
This exposes some of the "bosmang" pages for self-hosted users. This was
originally intended just for some quick things on goatcounter.com, but
it's useful enough to expose here too.

The sites overview isn't exposed, as it's mostly to set billing info and
such.

I should probably rename "bosmang" to "superuser" everywhere.

Only for users with access=superuser; sites which have only one user are
automatically upgraded.

Fixes #544