Update zdb; change -db syntax
This updated zdb, which includes quite a few internal improvements and
such.

It also changes the connection syntax from "sqlite://dbfile" to
"sqlite+dbfile"; especially with PostgreSQL's "postgres://..."-style
connection strings this caused a lot of confusion, basically since day
1.

To ease migration, we do a check if the string contains "://" but NOT
"+" and rewrite it and issue a warning. This should work for most cases,
except when people have a "+" in their password or something. The amount
of people self-hosting with PostgreSQL is probably not that many, and
the number of those using a "+" in their password (or even using a
password at all) even fewer, so it's probably fine.