"Lazy load" the dashboard
Previously the entire dashboard would be loaded when someone requested
it: every "widget" would run its query in a goroutine, and everything
runs in parallel.

This works fairly well, and I like that you get either a full dashboard
or nothing at all, and for most people the dashboard loads in <1s (often
much faster), but for some of the largest sites this is causing
problems:

1. Getting the data can take a few seconds.
2. Running several queries (10 now on the default dashboard) in parallel
   slows down the individual queries.
3. The dashboard is only as fast as the slowest widget.

To improve this, we now "lazy load" the dashboard a bit more:

1. On the first request, we still load the "internal widgets" and the
   first widget like before, and render that immediately. We could "lazy
   load" everything, but I don't really like how that looks.
2. Start getting the data for the other widgets, and send a "request ID"
   to the frontend.
3. Push this out over the websocket once it's done.

This is a nice balance between "lazy load everything" and "wait for
everything to finish rendering on the server". The problem with lazy
loading everything is that for a lot of cases it *seems* slower as
rendering the HTML, parsing the JS, establishing the websocket takes a
bit, sending data over it, and rendering that all takes a bit of time.
For the cases where the dashboard already renders fast (<200ms, which is
most cases) all of this decreases the perceived performance.

In the future we could also use the websocket to push updates/reloads
instead of doing a jQuery.Ajax request.