Make websocket connection global
Sometimes it wants to open multiple connections; I get error logs for
it, and everything seems to work alright since I haven't had any
complaints, but it shouldn't happen. I can't reproduce it either.

Make it a global and check the status. This is required anyway once we
start using it for more stuff.