Don't need Max
It would use a separate query/widget to get the maximum value for the
given time range; this is needed so we know how to scale the charts: the
second page may have a lot of visits for one day, even though the first
page has more visits overall.

But we can just get that by a for loop since we already have all that
data, we don't really need a query. Turns out that looping over 10 items
is a lot faster than scanning over 100k rows in SQL, who would have
guessed 'ey? This is a fairly significant performance win for larger
sites.

The reason it used to do a query is cases where a page has more visits
after pagination than the previous maximum value, but we can just
re-render the charts with the new max value; that used to be slow, but
now it's fast.