Don't lock tables for stat updates
I put this in in 78a7e8a, for PostgreSQL only. At the time this (crudly)
solved the problem of the 'goatcounter reindex' command conflicting with
the regular stat update process: it would try to update the same row
more than once and race.

At the time I was running 'reindex' somewhat frequently, but haven't for
a long time – the command has actually been removed.

The lock has stayed though, but I don't think there's actually any good
reason for that, and it's causing problems as PostgreSQL can't
autovacuum due to the full table lock.