Fix charts for SQLite
This fixes a regression introduces in 6bc0162; that changed so that stat
rows were no longer stored if there are no visits to store. That works
fine, except for hit_stats in SQLite since it will do a "select",
"delete", merges with new pageviews, and then an insert. But if there
are no new visits (only pageviews) it will never re-insert the rows.

So check the full array instead, and write a migration to correct the
wrong data.