Don't try to fetch a i18n locale if we don't need to
A z18n Locale is always added for every endpoint; this matters because
in some cases a translated message can appear even for endpoints where
it doesn't really matter (e.g. /count).

But we don't really need to do the whole matching logic for this. In
particular, we don't need to try to fetch the locale overrides for these
endpoints. This is unnecessary slow since it hits the database (simple
fast query, but still), and if the request is aborted by the client it
would result in a spurious "OverrideTranslations.Get: zdb.Get: context
canceled" message.

Fixes #637