Add "proxy" option in -tls
Give a hint that a secure connection will be used, so we know what value
to use for the cookie secure/samesite flags.

Maybe it would be better to auto-detect this per-connection though; that
would elimate the need for any flag. I don't see why we can't 🤔

Fixes #652
Also fixes part of #647