Revert "Allow URLs up to 8K, instead of 2K"
This reverts commit c0af8e110c7895ee3aa2e16a11c36100793d6ca5.

Turns out this doesn't actually work:

    pq: index row size 8000 exceeds btree version 4 maximum 2704 for index "paths#site_id#path"

The 2K limit is not really a huge deal, and it's more trouble than it's
worth to fix this (if at all possible: I didn't even look). No one has
complained about this, it's just some occasional pageviews that don't
work (only a few per day, and most for the same site).