Fix order for pages overview
The correct order was used inside the CTE, but *not* in the select from
the CTE. The order isn't guaranteed – it would nonetheless almost always
be the correct order, except when it wasn't.

All other CTEs were already correct except one for hit_stats.ByRef, so
fix that one too.

Fixes #552
Fixes #677