Use navigator.sendBeacon in count.js
The main advantage of this is that it's a lot shorter, and that it also
works with links to PDFs or other non-HTML content; using
<a href="file.pdf" data-goatcounter-event="file.pdf"> didn't work
because the request to /count didn't finish before the PDF started
loading, and the request would be aborted.

Still keep the img code as a fallback for now, mainly because "are
people still using very old browsers?" is a useful use case. Do keep
track of how many GET requests we get, so we can see if it's actually
used.