Don't send errors for Pikaday not being defined
e.g.:

    Jun 25 11:19:35 ERROR: JavaScript error: uncaught exception: ReferenceError: "Pikaday" is not defined. (https://static.zgo.at/jquery.js?v=fc4534d#2)
    {
        column="0"
        line="2"
        loc="https://XXXX/user/new"
        stack="@https://static.zgo.at/jquery.js?v=fc4534d:2\n"
        url="https://XXXX/user/new"
        userAgent="Mozilla/5.0 (Windows NT 6.1; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/83.0.4103.61 Safari/537.36"
    }

This always happens on /user/new, which is odd because that code doesn't
even run on that page, and always with the same old Chrome User-Agent
(probably some bot?) It's on a while bunch of different unrelated
domains though.

Not sure what's going on with this, and I can't reproduce it. Just
silence these errors.