Don't soft-delete sites; remove immediately
In all these years no one has asked me to undelete it, and it's caused
plenty of confusion because people delete sites to clear data, and then
re-create a new site with the same name.

So just delete it.

Still do the deletion in cron as it may take a while.