Remove "salt" from sessions
This doesn't really add anything, not really. The way this has worked
for a long time (since 2020, a1aa1c5) is:

1. A sessionHash is created as hash(salt + User-Agent + IP).
2. Store this in memory as a sessionHash→UUIDv4 map.
3. Use this UUIDv4 in the database and such.
4. The salt changes every few hours, resulting in new hashes and new
   UUIDs

So it never hits the disk, and the only thing the salt really adds is
rotation of the sessionHashes. This is different from before, where the
session ID was stored directly, but that was a long time ago and in the
very early versions only. So now it just expire the sessionHash after a
few hours.