Don't store individual pageviews by default
The pageviews in the hits table is never used for displaying the
dashboard, only for the CSV export. Not storing it has two advantages:

- More privacy-friendly; we only store aggregate data, not exact data.

- Uses less disk space, potentially *a lot* less for larger sites.

Disadvantage is that we can't rebuild the stats tables from the base
data, but that hasn't been done in years.

It also makes debugging a bit harder in some cases, but that's also
rare, and in those cases people can still enable it. Bots are always
stored for now, to make debugging easier if need be. They're
automatically deleted after 30 days though.

Another downside is that exporting pageviews won't be possible, because
this data will no longer exist. For most people: that's okay. They
either never export, or they can use the API for the dashboard stats.
For the rest: they can still enable it.