Replace ratelimiter with github.com/sethvargo/go-limiter
The problem with the zhttp ratelimiter is that the export limits
sometimes get "stuck". I'm not entirely sure why this happens, and only
on this one endpoint. It works on all the other endpoints, and the code
is identical to what we used at $job years ago.

I recently used go-limiter on another project, and that seems to work
well, so just use that – one less thing for me to maintain.

Fixes #748
Fixes #686
Fixes #674