Change default port to :8080, and don't use TLS/ACME by default
Given that the default of "tls,rdr,acme" was broken for some time
(585820e), I think it's safe to say this is not commonly used.

So change the default to just use plain HTTP on post :8080 by default.
The main advantage is that this is much easier to get started with, and
it seems *most* people use some sort of TLS proxy anyway.