Use correct text chart height when paginating
Because it did w.Max = 0, on pagination it would use the max of that
page, not taking the previous page in to account. That looked pretty
weird.

Not sure why it did that w.Max = 0; it's been this since the text
overview was added (d911a90c), so either something else changed since
then or it was always bugged.