Don't store in hits table for bots
It would always record bots in hits even if the "Individual pageviews"
was disabled, the idea being that it might be useful for debugging.

But I've never had to use since the default changed, and it's easier to
either always or never store it, especially for self-hosting. Can of
course still enable the setting if debugging is required.