Fix merging of pageviews
The old method of selecting all hits and just running them through
memstore no longer works, as that no longer exists for most people.

I looked at generating hits from the aggegrate stat tables, but that's
not so straight-forward. We can't just update the tables either, because
of the unique constraints.

So abstract all of this a little bit so we can do what cron does in
other places.

Fixes #801